backend/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==3.2
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
lxml==4.9.1
Pillow==9.2.0
PyJWT==2.4.0
pytz==2022.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.10
